rag-backend/pyproject.toml
[tool.poetry]
name = "rag-backend"
version = "0.1.0.rc1"
description = "You know, for shit."
authors = ["yongwoo <yongwoo@crowdworks.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
aiosqlite = "0.19.0"
unstructured = {python = ">=3.9.0,<3.12", version = "^0.12.4"}
aiomysql = "0.2.0"
mysql-connector-python = "8.3.0"
python-multipart = "0.0.9"
fastapi = "^0.109.2"
prometheus-client = "^0.20.0"
prometheus-fastapi-instrumentator = "^6.1.0"
tqdm = "^4.66.1"
uvicorn = "^0.27.0.post1"
starlette-context = "^0.3.6"
black = "^24.1.1"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
httpx = "^0.26.0"
gunicorn = "^21.2.0"
langchain = "^0.1.6"
chroma = "^0.2.0"
sentence-transformers = "^2.3.1"
chromadb = "^0.4.22"
bs4 = "^0.0.2"
pypdf = "^4.0.1"
markdown = "^3.5.2"
multipledispatch = "^1.0.0"
docx2txt = "^0.8"
langchain-extensions = {git = "https://bitbucket.org/crowdworks_dev/langchain-extensions.git"}
langchain-openai = "^0.0.6"
langchainhub = "^0.1.14"
pydantic = "^2.6.1"
streamlit = "^1.31.1"
grandalf = "^0.8"
openpyxl = "^3.1.2"

[tool.poetry.group.dev.dependencies]
mock = "^5.1.0"
pytest = "^8.0.0"
pytest-env = "^1.1.3"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.4"
pytest-freezegun = "^0.4.2"
pytest-xdist = "^3.5.0"
pre-commit = "^3.2.0"
pytest-cov = "^4.1.0"
locust = "^2.23.1"
watchdog = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
